+++ builder/src/main/java/com/stratio/cassandra/lucene/builder/index/schema/mapping/Mapper.java
@@ -15,10 +15,10 @@
  */
 package com.stratio.cassandra.lucene.builder.index.schema.mapping;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.stratio.cassandra.lucene.builder.JSONBuilder;
-import org.codehaus.jackson.annotate.JsonProperty;
-import org.codehaus.jackson.annotate.JsonSubTypes;
-import org.codehaus.jackson.annotate.JsonTypeInfo;
 
 /**
  * Class for mapping between Cassandra's columns and Lucene documents.